Improved: the logic for fixing infinite scroll(#84)
Removed the logic to re-render the infinite-scroll on queryString check.
Added a variable to check whether the scrolling can be enabled or not whenever users lands on the list page. For this, we have determined the height of the content part scroll and infiniteScroll component and if the height is less than 0 then only enabling the infinite-scroll component